Add basic filters and security to cardsController

+++ src/controllers/cardsController.js
@@ -32,11 +32,22 @@ export default async function cardsController(fastify) {
 
   // List cards
   fastify.get('/', async (request, reply) => {
-    return Card.find()
+    const { sub: userId } = request.user
+    const filter = { userId }
+    const { collection: collectionId, deck: deckIds } = request.query
+
+    if (deckIds) {
+      filter.deckIds = deckIds
+    } else if (collectionId) {
+      filter.collectionId = collectionId
+    }
+
+    return Card.find(filter)
   })
 
   // Add card
   fastify.post('/new', async (request, reply) => {
+    const { sub: userId } = request.user
     const { deckIds } = request.body
     const newCard = new Card(request.body)
 
@@ -45,16 +56,25 @@ export default async function cardsController(fastify) {
 
   // Show one card
   fastify.get('/:id', async (request, reply) => {
-    return Card.findOne({ id: request.params.id })
+    const { sub: userId } = request.user
+
+    return Card.findOne({ id: request.params.id, userId })
   })
 
   // Edit card
   fastify.put('/:id', async (request, reply) => {
-    return Card.findOneAndUpdate({ id: request.params.id }, request.body)
+    const { sub: userId } = request.user
+
+    return Card.findOneAndUpdate(
+      { id: request.params.id, userId },
+      request.body,
+    )
   })
 
   // Delete card
   fastify.delete('/:id', async (request, reply) => {
-    return Card.findOneAndRemove({ id: request.params.id })
+    const { sub: userId } = request.user
+
+    return Card.findOneAndRemove({ id: request.params.id, userId })
   })
 }
